+++ src/CControl/Sources/LinearAlgebra/expm.c
@@ -12,7 +12,7 @@
  * A[m*n]
  * m == n
  */
-void expm(float A[], uint32_t row){
+void expm(float A[], uint16_t row){
 	// Create zero matrix
 	float E[row*row];
 	memset(E, 0, sizeof(E));
@@ -21,23 +21,23 @@ void expm(float A[], uint32_t row){
 	float T[row*row];
 	memset(F, 0, sizeof(F));
 	memset(T, 0, sizeof(T));
-	for(uint32_t i = 0; i < row; i++){
+	for(uint16_t i = 0; i < row; i++){
 		F[i*row + i] = 1;
 		T[i*row + i] = 1;
 	}
 	uint8_t k = 1;
 	while (norm(T, row, row, 1) > 0){
 		// E = E + F
-		for(uint32_t i = 0; i < row*row; i++){
+		for(uint16_t i = 0; i < row*row; i++){
 			E[i] = E[i] + F[i];
 		}
 		// F = A*F/k (we are borrowing T)
 		mul(A, F, T, row, row, row);
-		for(uint32_t i = 0; i < row*row; i++){
+		for(uint16_t i = 0; i < row*row; i++){
 			F[i] = T[i]/k;
 		}
 		// T = E + F - E ( This is faster than T = F)
-		for(uint32_t i = 0; i < row*row; i++){
+		for(uint16_t i = 0; i < row*row; i++){
 			T[i] = E[i] + F[i] - E[i];
 		}
 		k++;
